requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pyyaml>=6.0.1
packaging>=24.0
